requirements.txt
aiogram==2.21
aiohttp==3.8.1
aioschedule==0.5.2
aiosignal==1.2.0
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
Babel==2.9.1
bestchange-api==2.0.0.0
certifi==2021.10.8
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
commonmark==0.9.1
frozenlist==1.3.1
h11==0.13.0
idna==3.3
multidict==6.0.2
outcome==1.2.0
prompt-toolkit==3.0.30
pydantic==1.9.2
Pygments==2.13.0
pyngrok==5.1.0
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
rich==12.5.1
shellingham==1.5.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.12
wcwidth==0.2.5
websockets==10.3
websockets-cli==0.2.0
wsproto==1.1.0
yarl==1.8.1
